+++ NEWS.md
@@ -1,5 +1,9 @@
 # nanonext 0.6.0.9200 (development)
 
+#### New Features
+
+* `status_code()` utility included for returning a human-readable translation of HTTP response status codes.
+
 #### Updates
 
 *Please review the following potentially breaking changes, and only update when ready:*
@@ -13,7 +17,6 @@
 *Other changes:*
 
 * Fixes bug introduced in 0.6.0 where Aios returning 'errorValues' are not cached with the class, returning only integer values when accessed subsequently.
-* `ncurl()` status codes other than 200 (success) now contain a character attribute 'status' with an explanation of the meaning.
 * `device()` no longer prompts for confirmation in interactive environments - as device creation is only successful when binding 2 raw mode sockets, there is little scope for accidental use.
 * Print method for 'errorValue' now also provides the human translation of the error code.
 * Bundled 'libnng' source updated to v1.6.0 pre-release (5385b78).